+++ test/records/test_get_all_payments.js
@@ -1,4 +1,4 @@
-const {test} = require('tap');
+const {test} = require('@alexbosworth/tap');
 
 const {getAllPayments} = require('./../../records');